Implementation of Lane class and collision detection
Implemented the main logic of Lane class, that stores a list of obstacles, powerups and boats. Created methods to access and modify its correcponding attributes and the collision detection logic(using collision handler yet to implement).
Created Collidable interface and CollidableVisitor to facilitate the collision detection to the Lane class. Modified obsctacle classes to implement collidable interface and accept ColidabelVisitors.