Implemented Observer pattern for playerControlledStrategy class and modularity with the keybinds
Created UserAction enum, that lists all the posible things a user can do with the keyboard.
Created KeyBindings class, that maps the Keycodes with the UserActions.
Created InputSubscribed Interface for observers of the Input events, this interface is implemented by
PlayerControlledStrategy, to be notified of the actions to perform by the InputManager class (helper clas of LegScreen)