Implented PlayerControlledStrategy class correctly
Corrected Observer pattern, now Player is subscribed to Inputs of the InputManager, and Player notifies PlayerControlledStrategy when required.
Modifications in class Boat and GameObject, now boats are created with default posiiton (0,0) to be changed by the lanes, and GameObject stores boolean wasHit and implements getter and setter for all the Collidables.