Mayor refactoring: change responsability level of movement and creation of GameObjects. Functional configuration of AIControlledStrategy
Now the movement is handled by the Leg class instead than by the Lane class. The gameObjects for each lane are created in the Leg class, and passed to the each of the Lanes in their creation.
The collision responsability remains in the Lanes.
Created ObstacleType enum and PowerUpType enum, added factory methods to both classes to better handle the creation. Impelmented a randomType creation method for both Obstacles and PowerUps (In Boats was already implemented).
Documentation of all the mathods and changes (I hope)
Minor changes (basic AIMovementStrategy impentation, whiteline erased, other format changes...)
Modification of factory method in class Boat, now it doen't receive a boolean with the role, but its moveStrategy has to be assigned after its creation.
In Leg class, when the alens are created, if the lane is the one of a rival, the leg passes to the moveStrategy a reference to the lane in which the boat is.